Fixes test Renames classes to match the interfaces they implement
