+++ Python/Product/PythonTools/PythonTools/WebProject/PythonWebLauncher.cs
@@ -16,27 +16,14 @@
 
 using System;
 using System.Collections.Generic;
-using System.ComponentModel.Composition;
 using System.Diagnostics;
-using System.Globalization;
-using System.IO;
 using System.Linq;
 using System.Net.NetworkInformation;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading;
 using System.Threading.Tasks;
-using System.Web;
-using System.Windows.Forms;
 using Microsoft.PythonTools.Debugger;
-using Microsoft.PythonTools.Debugger.DebugEngine;
 using Microsoft.PythonTools.Infrastructure;
-using Microsoft.PythonTools.Intellisense;
 using Microsoft.PythonTools.Interpreter;
 using Microsoft.VisualStudio;
-using Microsoft.VisualStudio.Debugger;
-using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudioTools;
 using Microsoft.VisualStudioTools.Project;
@@ -50,8 +37,6 @@ namespace Microsoft.PythonTools.Project.Web {
     /// appropriate page into the users web browser.
     /// </summary>
     class PythonWebLauncher : IProjectLauncher {
-        private int? _testServerPort;
-
         public const string RunWebServerCommand = "PythonRunWebServerCommand";
         public const string DebugWebServerCommand = "PythonDebugWebServerCommand";
 
@@ -96,7 +81,7 @@ namespace Microsoft.PythonTools.Project.Web {
 
             Uri url;
             int port;
-            GetFullUrl(config, out url, out port);
+            GetFullUrl(_serviceProvider, config, out url, out port);
 
             var env = new Dictionary<string, string> { { "SERVER_PORT", port.ToString() } };
             if (url != null) {
@@ -108,6 +93,7 @@ namespace Microsoft.PythonTools.Project.Web {
             if (debug) {
                 _pyService.Logger.LogEvent(Logging.PythonLogEvent.Launch, 1);
 
+                config.LaunchOptions[PythonConstants.WebBrowserUrlSetting] = url.AbsoluteUri;
                 using (var dsi = DebugLaunchHelper.CreateDebugTargetInfo(_serviceProvider, config)) {
                     dsi.Launch();
                 }
@@ -135,7 +121,7 @@ namespace Microsoft.PythonTools.Project.Web {
 
         private Task StartBrowser(string url, Func<bool> shortCircuitPredicate) {
             Uri uri;
-            if (!String.IsNullOrWhiteSpace(url) && Uri.TryCreate(url, UriKind.RelativeOrAbsolute, out uri)) {
+            if (!string.IsNullOrWhiteSpace(url) && Uri.TryCreate(url, UriKind.RelativeOrAbsolute, out uri)) {
                 var tcs = new TaskCompletionSource<object>();
 
                 OnPortOpenedHandler.CreateHandler(
@@ -173,7 +159,7 @@ namespace Microsoft.PythonTools.Project.Web {
 
         #endregion
 
-        private void GetFullUrl(LaunchConfiguration config, out Uri uri, out int port) {
+        internal static void GetFullUrl(IServiceProvider provider, LaunchConfiguration config, out Uri uri, out int port) {
             var host = config.GetLaunchOption(PythonConstants.WebBrowserUrlSetting);
             if (string.IsNullOrEmpty(host)) {
                 uri = null;
@@ -196,13 +182,16 @@ namespace Microsoft.PythonTools.Project.Web {
                 if (int.TryParse(config.GetLaunchOption(PythonConstants.WebBrowserPortSetting) ?? "", out p)) {
                     builder.Port = p;
                 } else if (builder.Port < 0 || (builder.Uri.IsDefaultPort && !host.Contains(":{0}".FormatInvariant(builder.Port)))) {
-                    builder.Port = TestServerPort;
+                    builder.Port = GetTestServerPort();
                 }
 
                 uri = builder.Uri;
                 port = uri.Port;
             } catch (UriFormatException) {
-                var output = OutputWindowRedirector.GetGeneral(_serviceProvider);
+                if (provider == null) {
+                    throw;
+                }
+                var output = OutputWindowRedirector.GetGeneral(provider);
                 output.WriteErrorLine(Strings.ErrorInvalidLaunchUrl.FormatUI(host));
                 output.ShowAndActivate();
                 uri = null;
@@ -210,15 +199,10 @@ namespace Microsoft.PythonTools.Project.Web {
             }
         }
 
-        private int TestServerPort {
-            get {
-                if (!_testServerPort.HasValue) {
-                    _testServerPort = Enumerable.Range(new Random().Next(49152, 65536), 60000)
-                        .Except(IPGlobalProperties.GetIPGlobalProperties().GetActiveTcpConnections().Select(c => c.LocalEndPoint.Port))
-                        .First();
-                }
-                return _testServerPort.Value;
-            }
+        private static int GetTestServerPort() {
+            return Enumerable.Range(new Random().Next(49152, 65536), 60000)
+                .Except(IPGlobalProperties.GetIPGlobalProperties().GetActiveTcpConnections().Select(c => c.LocalEndPoint.Port))
+                .First();
         }
     }
 }
