+++ Python/Product/Workspace/Workspace.csproj
@@ -124,10 +124,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="PyprojDebugLaunchProvider.cs" />
     <Compile Include="PyprojLaunchDebugTargetProvider.cs" />
+    <Compile Include="PyprojLaunchConfigurationProvider.cs" />
+    <Compile Include="PythonLaunchConfigurationProvider.cs" />
     <Compile Include="PythonLaunchDebugTargetProvider.cs" />
-    <Compile Include="PythonDebugLaunchProvider.cs" />
   </ItemGroup>
   <Import Project="..\ProjectAfter.settings" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
